chore: moved partner asset to separate script
