+++ Runtime/AvatarCreator/Data/AssetLibrary.cs
@@ -1,25 +1,9 @@
-﻿using Newtonsoft.Json;
-using ReadyPlayerMe.Core;
-
-namespace ReadyPlayerMe.AvatarCreator
+﻿namespace ReadyPlayerMe.AvatarCreator
 {
     public struct AssetLibrary
     {
         public PartnerAsset[] Assets;
         public Pagination Pagination;
-    }
-
-    public struct PartnerAsset
-    {
-        public string Id;
-        [JsonProperty("type"), JsonConverter(typeof(CategoryConverter))]
-        public Category Category;
-        [JsonConverter(typeof(GenderConverter))]
-        public OutfitGender Gender;
-        [JsonProperty("iconUrl")]
-        public string ImageUrl;
-        [JsonProperty("lockedCategories")]
-        public string[] LockedCategories;
     }
 
     public struct Pagination
