+++ Runtime/AvatarCreator/Data/PartnerAsset.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f2719f45a60bfd443829b433870e2b2f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
